moving stateProofMessage from crypto/compcert to compcert
