+++ compactcert/msgp_gen_test.go
@@ -12,6 +12,66 @@ import (
 	"github.com/algorand/msgp/msgp"
 )
 
+func TestMarshalUnmarshalStateProofMessage(t *testing.T) {
+	partitiontest.PartitionTest(t)
+	v := StateProofMessage{}
+	bts := v.MarshalMsg(nil)
+	left, err := v.UnmarshalMsg(bts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(left) > 0 {
+		t.Errorf("%d bytes left over after UnmarshalMsg(): %q", len(left), left)
+	}
+
+	left, err = msgp.Skip(bts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(left) > 0 {
+		t.Errorf("%d bytes left over after Skip(): %q", len(left), left)
+	}
+}
+
+func TestRandomizedEncodingStateProofMessage(t *testing.T) {
+	protocol.RunEncodingTest(t, &StateProofMessage{})
+}
+
+func BenchmarkMarshalMsgStateProofMessage(b *testing.B) {
+	v := StateProofMessage{}
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		v.MarshalMsg(nil)
+	}
+}
+
+func BenchmarkAppendMsgStateProofMessage(b *testing.B) {
+	v := StateProofMessage{}
+	bts := make([]byte, 0, v.Msgsize())
+	bts = v.MarshalMsg(bts[0:0])
+	b.SetBytes(int64(len(bts)))
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		bts = v.MarshalMsg(bts[0:0])
+	}
+}
+
+func BenchmarkUnmarshalStateProofMessage(b *testing.B) {
+	v := StateProofMessage{}
+	bts := v.MarshalMsg(nil)
+	b.ReportAllocs()
+	b.SetBytes(int64(len(bts)))
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, err := v.UnmarshalMsg(bts)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
 func TestMarshalUnmarshalsigFromAddr(t *testing.T) {
 	partitiontest.PartitionTest(t)
 	v := sigFromAddr{}
