+++ crypto/compactcert/verifier.go
@@ -56,7 +56,7 @@ func (v *Verifier) Verify(c *Cert) error {
 	sigs := make(map[uint64]crypto.Hashable)
 	parts := make(map[uint64]crypto.Hashable)
 
-	msghash := v.Params.StateProofMessage.Hash()
+	msghash := v.Params.StateProofMessageHash
 	for pos, r := range c.Reveals {
 		sig, err := buildCommittableSignature(r.SigSlot)
 		if err != nil {
